+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # CHANGELOG.md
 
-## Unreleased
+## Unreleased v0.1
 
 ### Added
 - Filesystem utilities.
 - Project-specific exception hierarchy.
 - Audio metadata reader.
+
+## Unreleased v0.2.1
+
+### Added
+- Gen1 library loader.
+- Audio metadata reader.
+- Automatic logo detection.
+- Single-file station support.